重构：作为库使用时，尽量避免 panic 的使用，所有错误返回 error

+++ render/render.go
@@ -30,25 +30,31 @@ func StreamingRender(output io.Writer, format string, noHeader bool, colNames []
 }
 
 func streamingRender(output io.Writer, format string, stream <-chan map[string]interface{}, noHeader bool, colNames []string) (int, error) {
-	if format == "csv" {
-		// Write BOM header for UTF-8
-		must.Must(output.Write([]byte("\xEF\xBB\xBF")))
-	}
-
 	var total int
 	switch format {
 	case "json":
 		for item := range stream {
 			total++
-			must.Must(output.Write(must.Must(json.Marshal(item))))
-			must.Must(output.Write([]byte("\n")))
+			if _, err := output.Write(must.Must(json.Marshal(item))); err != nil {
+				return 0, err
+			}
+			if _, err := output.Write([]byte("\n")); err != nil {
+				return 0, err
+			}
 		}
 	case "csv":
+		// Write BOM header for UTF-8
+		if _, err := output.Write([]byte("\xEF\xBB\xBF")); err != nil {
+			return 0, err
+		}
+
 		csvWriter := csv.NewWriter(output)
 		defer csvWriter.Flush()
 
 		if !noHeader {
-			must.NoError(csvWriter.Write(colNames))
+			if err := csvWriter.Write(colNames); err != nil {
+				return 0, err
+			}
 		}
 
 		for item := range stream {
@@ -58,7 +64,9 @@ func streamingRender(output io.Writer, format string, stream <-chan map[string]i
 				line = append(line, resolveValue(item[colName]))
 			}
 
-			must.NoError(csvWriter.Write(line))
+			if err := csvWriter.Write(line); err != nil {
+				return 0, err
+			}
 		}
 	default:
 		for item := range stream {
@@ -75,23 +83,44 @@ func streamingRender(output io.Writer, format string, stream <-chan map[string]i
 				)
 			}
 
-			must.Must(output.Write([]byte(fmt.Sprintln(strings.Join(lines, ", ")))))
+			if _, err := output.Write([]byte(fmt.Sprintln(strings.Join(lines, ", ")))); err != nil {
+				return 0, err
+			}
 		}
 	}
 
 	return total, nil
 }
 
-func streamRenderXlsx(output io.Writer, noHeader bool, colNames []string, stream <-chan map[string]interface{}) (int, error) {
+func streamRenderXlsx(output io.Writer, noHeader bool, colNames []string, stream <-chan map[string]interface{}) (total int, err error) {
 	tmpFilename := createTempFilename() + ".xlsx"
-	w := must.Must(NewExcelWriter(tmpFilename, ternary.If(noHeader, []string{}, colNames)))
+	w, err := NewExcelWriter(tmpFilename, ternary.If(noHeader, []string{}, colNames))
+	if err != nil {
+		return 0, err
+	}
+
 	defer func() {
-		must.NoError(w.Close())
-		must.Must(io.Copy(output, must.Must(os.Open(tmpFilename))))
-		_ = os.Remove(tmpFilename)
+		if err1 := w.Close(); err1 != nil {
+			err = err1
+			return
+		}
+
+		f, err1 := os.Open(tmpFilename)
+		if err1 != nil {
+			err = err1
+			return
+		}
+		defer func() {
+			_ = f.Close()
+			_ = os.Remove(tmpFilename)
+		}()
+
+		if _, err1 = io.Copy(output, f); err1 != nil {
+			err = err1
+			return
+		}
 	}()
 
-	var total int
 	for item := range stream {
 		total++
 		line := make([]string, 0)
@@ -99,7 +128,9 @@ func streamRenderXlsx(output io.Writer, noHeader bool, colNames []string, stream
 			line = append(line, resolveValue(item[colName]))
 		}
 
-		must.NoError(w.Write(line))
+		if err := w.Write(line); err != nil {
+			return 0, err
+		}
 	}
 
 	return total, nil
@@ -113,26 +144,26 @@ func resolveValue(value interface{}) string {
 	return fmt.Sprintf("%v", value)
 }
 
-func Render(format string, noHeader bool, colNames []string, kvs []map[string]interface{}, sqlStr string) *bytes.Buffer {
+func Render(format string, noHeader bool, colNames []string, kvs []map[string]interface{}, sqlStr string) (*bytes.Buffer, error) {
 	writer := bytes.NewBuffer(nil)
 
 	switch format {
 	case "json":
-		JSON(writer, kvs)
+		return writer, JSON(writer, kvs)
 	case "yaml":
-		YAML(writer, kvs)
+		return writer, YAML(writer, kvs)
 	case "table":
-		Table(writer, noHeader, colNames, kvs)
+		return writer, Table(writer, noHeader, colNames, kvs)
 	case "markdown":
-		Markdown(writer, noHeader, colNames, kvs)
+		return writer, Markdown(writer, noHeader, colNames, kvs)
 	case "csv":
-		CSV(writer, noHeader, colNames, kvs)
+		return writer, CSV(writer, noHeader, colNames, kvs)
 	case "html":
-		HTML(writer, noHeader, colNames, kvs)
+		return writer, HTML(writer, noHeader, colNames, kvs)
 	case "xlsx":
-		XLSX(writer, noHeader, colNames, kvs)
+		return writer, XLSX(writer, noHeader, colNames, kvs)
 	case "xml":
-		XML(writer, colNames, kvs, sqlStr)
+		return writer, XML(writer, colNames, kvs, sqlStr)
 	default:
 		for _, kv := range kvs {
 			lines := make([]string, 0)
@@ -144,11 +175,14 @@ func Render(format string, noHeader bool, colNames []string, kvs []map[string]in
 				), "\n", "\\n"))
 			}
 
-			writer.Write([]byte(fmt.Sprintln(strings.Join(lines, ", "))))
+			_, err := writer.Write([]byte(fmt.Sprintln(strings.Join(lines, ", "))))
+			if err != nil {
+				return nil, err
+			}
 		}
 	}
 
-	return writer
+	return writer, nil
 }
 
 func createTempFilename() string {
